[rpc][api] TestPlan.{get,check}_plan_type -> PlanType.filter
